requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.0
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mega.py==1.0.8
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
passlib==1.7.4
pathlib==1.0.1
pillow==10.4.0
psycopg2==2.9.9
pyasn1==0.6.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
